Keep scrape config in line with the new Prometheus scrape config
This is triggered by https://github.com/grafana/jsonnet-libs/pull/261 .

The above PR changes the `instance` label to be actually unique within
a scrape config. It also adds a `pod` and a `container` target label
so that metrics can easily be joined with metrics from cAdvisor, KSM,
and the Kubelet.

This commit adds the same to the Loki scrape config. It also removes
the `container_name` label. It is the same as the `container` label
and was already added to Loki previously. However, the
`container_name` label is deprecated and has disappeared in K8s 1.16,
so that it will soon become useless for direct joining.